pyproject.toml
[project]
name = "django-sqlagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "amqp==5.3.1",
    "annotated-types==0.7.0",
    "anthropic==0.84.0",
    "anyio==4.12.1",
    "asgiref==3.11.1",
    "attrs==25.4.0",
    "billiard==4.2.4",
    "brotli==1.2.0",
    "celery==5.6.2",
    "certifi==2026.1.4",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1.2",
    "click-repl==0.3.0",
    "colorama==0.4.6",
    "cryptography==46.0.5",
    "dataclasses-json==0.6.7",
    "ddgs==9.10.0",
    "distro==1.9.0",
    "dj-database-url==3.1.2",
    "dj-rest-auth==7.1.1",
    "django==6.0.2",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "docstring-parser==0.17.0",
    "duckduckgo-search==8.1.1",
    "dydantic==0.0.8",
    "fake-useragent==2.2.0",
    "filetype==1.2.0",
    "frozenlist==1.8.0",
    "google-auth==2.48.0",
    "google-genai==1.64.0",
    "greenlet==3.3.2",
    "groq==0.37.1",
    "h11==0.16.0",
    "h2==4.3.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "hyperframe==6.1.0",
    "idna==3.11",
    "jiter==0.13.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "kombu==5.6.2",
    "langchain==1.2.10",
    "langchain-anthropic==1.3.4",
    "langchain-classic==1.0.1",
    "langchain-community==0.4.1",
    "langchain-core==1.2.16",
    "langchain-google-genai==4.2.1",
    "langchain-groq==1.1.2",
    "langchain-openai==1.1.10",
    "langchain-text-splitters==1.1.1",
    "langgraph==1.0.9",
    "langgraph-checkpoint==4.0.0",
    "langgraph-checkpoint-postgres==3.0.4",
    "langgraph-prebuilt==1.0.8",
    "langgraph-sdk==0.3.8",
    "langmem==0.0.30",
    "langsmith==0.7.6",
    "lxml==6.0.2",
    "marshmallow==3.26.2",
    "multidict==6.7.1",
    "mypy-extensions==1.1.0",
    "numpy==2.4.2",
    "openai==2.24.0",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "packaging==26.0",
    "primp==1.0.0",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "psycopg==3.3.3",
    "psycopg-binary==3.3.3",
    "psycopg-pool==3.3.0",
    "psycopg2-binary==2.9.11",
    "pyasn1==0.6.2",
    "pyasn1-modules==0.4.2",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.13.1",
    "pyjwt==2.11.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "redis==7.2.0",
    "regex==2026.2.28",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "socksio==1.0.0",
    "sqlalchemy==2.0.46",
    "sqlparse==0.5.5",
    "tenacity==9.1.4",
    "tiktoken==0.12.0",
    "tqdm==4.67.3",
    "trustcall==0.0.39",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "tzlocal==5.3.1",
    "urllib3==2.6.3",
    "uuid-utils==0.14.1",
    "vine==5.1.0",
    "wcwidth==0.6.0",
    "websockets==15.0.1",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zstandard==0.25.0",
]
